temporary remove AsynEngine usage from Median Filter, because both non completed yet

+++ PREMIER_PLUGIN.DEV/ImageLabFuzzyMedian/include/ImageLabFuzzyMedian.h
@@ -56,7 +56,7 @@ static_assert((MinKernelWidth & 0x1), "Kernel width value must be ODD");
 constexpr int MaxKernelWidth  = kernel_width(MaxKernelRadius);
 static_assert((MaxKernelWidth & 0x1), "Kernel width value must be ODD");
 
-typedef	uint16_t	HistElem;
+typedef	uint32_t	HistElem;
 constexpr csSDK_int32 sizeOfHistElem = static_cast<csSDK_int32>(sizeof(HistElem));
 
 constexpr csSDK_int32 used_mem_size = CreateAlignment(512 * 1024, CPU_PAGE_SIZE);
